Updated main.py and created UI.py
In main.py, I added the drop command to remove certain columns from my dataframe. When I adjusted my code, it only had the required columns that I need, which is on my Data Dictionary. The code worked, as it removed some unnecessary columns. In UI.py, I started making the User Interface, by trying to make a definition for "most_delayed_airline". I tried multiple ways to drop rows based on what it says on the dataframe, but it didn't work. Next, I need to figure out how to do it and make more definitions for the different categories.