Update main.py and UI.py
In main.py, I changed it so that .csv file only shows the delayed flights. After a lot of testing, it worked, and was successful. in UI.py, is started making the different categories for which the data will be displayed, like "most_delayed_airline" and "most_delayed_time_block". Next, I need to continue this and finish the user interface.