Updated main.py, and UI.py
In main.py, I changed some of the columns that I wanted dropped as it wasn't necessary in the final result. I changed one new column as I made a mistake in it before. In UI.py, I defined all of the functions which I used in the user interface. Then, I made the user interface with lots of options to choose. Next, I need to fix it because there are some problems with it.