Update UI.py
This lesson, I spent most of the time trying to make the y axis in my matplotlib chart in order, but with multiple attempts and with the help of Mr Scott, it sadly didn't work. There are still some errors in the user interface so that is something that I need to work on, but other than that everything else works well.