Update main.py and UI.py
In main.py, I changed each value for the month so that instead of "1" it says January, and the same for the rest of the months. I also did this for the days of the week. I also made those into a string. In UI.py,  I added .values[0] to get rid of the 0 that kept showing before the answer. I added the airport because I forgot to do that before, and made the other functions that were floats into string. I also made the most_delayed_TMAX_CELSIUS to 2 decimal places as it looked neater. Next I need to make the READ.ME and do the evaluations.